Add a dedicated class for MANAGE tests.
